requirements.txt
agno==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.0
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
curl-cffi==0.13.0
cycler==0.12.1
ddgs==9.10.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
fake-useragent==2.2.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
feedparser==6.0.12
fonttools==4.61.1
frozendict==2.4.7
gitdb==4.0.12
gitpython==3.1.46
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
multitasking==0.0.12
numpy==2.4.1
ollama==0.6.1
openai==2.14.0
packaging==25.0
pandas==3.0.0
patsy==1.0.2
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
portalocker==3.2.0
primp==0.15.0
protobuf==6.33.4
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.3.2
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
pyyaml==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
scipy==1.17.0
seaborn==0.13.2
sentry-sdk==2.50.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
sqlalchemy==2.0.45
sse-starlette==3.2.0
starlette==0.50.0
statsmodels==0.14.6
tqdm==4.67.1
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
yfinance==1.0
